Conroller Fix
Fixed PlayController to work like home controller. Added dark mode throughout the app
