+++ NewPickFlicks/NewPickFlicks.xcodeproj/xcuserdata/paulbunnell.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -7,7 +7,7 @@
 		<key>NewPickFlicks.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>36</integer>
+			<integer>33</integer>
 		</dict>
 	</dict>
 </dict>
