+++ NewPickFlicks/NewPickFlicks/Controller/MainTabController.swift
@@ -50,7 +50,7 @@ class MainTabController: UITabBarController {
     func configureViewController(withUser user: User) {
         view.backgroundColor = .white
 
-        let home = templateNAvigationController(unselectedImage: #imageLiteral(resourceName: "icons8-movie-ticket-50"), selectedImage: #imageLiteral(resourceName: "icons8-movie-ticket-50-2"), rootViewController: HomeController(user: user))
+        let home = templateNAvigationController(unselectedImage: #imageLiteral(resourceName: "icons8-movie-ticket-50"), selectedImage: #imageLiteral(resourceName: "icons8-movie-ticket-50-2"), rootViewController: PlayController(user: user))
         
         let search = templateNAvigationController(unselectedImage: #imageLiteral(resourceName: "Search"), selectedImage: #imageLiteral(resourceName: "Search"), rootViewController: SearchController(config: .all))
